API updated, http tests added

+++ IE/src/main/java/ru/kpfu/itis/belskaya/controllers/OrderRestController.java
@@ -4,29 +4,24 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.TypeDescriptor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
-import ru.kpfu.itis.belskaya.converters.OrderFormToOrderConverter;
+import ru.kpfu.itis.belskaya.controllers.api.OrderApi;
+import ru.kpfu.itis.belskaya.converters.OrderConverter;
 import ru.kpfu.itis.belskaya.models.*;
-import ru.kpfu.itis.belskaya.models.forms.OrderForm;
-import ru.kpfu.itis.belskaya.services.AccountService;
+import ru.kpfu.itis.belskaya.models.dto.OrderDto;
 import ru.kpfu.itis.belskaya.services.OrderService;
 import ru.kpfu.itis.belskaya.services.SubjectService;
 import ru.kpfu.itis.belskaya.services.UserService;
 
+import javax.validation.Valid;
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Controller
 @RestController
-@RequestMapping("/api")
-public class OrderRestController {
-
-    @Autowired
-    private AccountService accountService;
+public class OrderRestController implements OrderApi {
 
     @Autowired
     private OrderService orderService;
@@ -40,37 +35,56 @@ public class OrderRestController {
     @Autowired
     private SubjectService subjectService;
 
-    @DeleteMapping("/{id}")
-    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<String> deleteOrder(@PathVariable("id") Long id) {
+    public ResponseEntity<OrderDto> getOrder(Long id) {
         Order order = orderService.findOrderById(id).orElseThrow(() -> new ResponseStatusException(
-                HttpStatus.NOT_FOUND, "Book not found"
+                HttpStatus.NOT_FOUND, "Order not found"
         ));
+
+        OrderConverter orderConverter = new OrderConverter(null);
+        OrderDto orderDto = (OrderDto) orderConverter.convert(order, TypeDescriptor.valueOf(Order.class), TypeDescriptor.valueOf(OrderDto.class));
+        return ResponseEntity.ok(orderDto);
+    }
+
+    @Override
+    public ResponseEntity<String> addOrder(Account account, OrderDto orderDto) {
+        Student student = studentService.findByAccount_Id(account.getId());
+        OrderConverter converter = new OrderConverter(student);
+        Order updatedOrder = (Order) converter.convert(orderDto, TypeDescriptor.valueOf(OrderDto.class), TypeDescriptor.valueOf(Order.class));
+        updatedOrder.setTutor(null);
+        orderService.updateOrder(updatedOrder);
+        return ResponseEntity.accepted().build();
+    }
+
+    public ResponseEntity<String> deleteOrder(Account account, Long id) {
+        Order order = orderService.findOrderById(id).orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Order not found"
+        ));
+        if (order.getAuthor().getAccount().equals(account)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You have no rights to delete this order");
+        }
         orderService.deleteOrder(order);
         return ResponseEntity.ok("Deleted");
     }
 
-    @PutMapping("/{id}")
-    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<String> updateOrder(@PathVariable("id") Long id,
-                                              @RequestBody OrderForm orderForm,
-                                              @AuthenticationPrincipal Account account) {
+    public ResponseEntity<String> updateOrder(Account account, Long id, @Valid OrderDto orderDto) {
         Student student = studentService.findByAccount_Id(account.getId());
-        OrderFormToOrderConverter converter = new OrderFormToOrderConverter(student);
-        Order updatedOrder = (Order) converter.convert(orderForm, TypeDescriptor.valueOf(OrderForm.class), TypeDescriptor.valueOf(Order.class));
+        OrderConverter converter = new OrderConverter(student);
+        Order updatedOrder = (Order) converter.convert(orderDto, TypeDescriptor.valueOf(OrderDto.class), TypeDescriptor.valueOf(Order.class));
         Order orderById = orderService.findOrderById(id).orElseThrow(() -> new ResponseStatusException(
                 HttpStatus.NOT_FOUND, "Order not found"
         ));
+        if (orderById.getAuthor().getAccount().equals(account)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You have no rights to delete this order");
+        }
         updatedOrder.setId(orderById.getId());
         updatedOrder.setCreationDate(orderById.getCreationDate());
+        updatedOrder.setCandidates(orderById.getCandidates());
+        updatedOrder.setTutor(orderById.getTutor());
         orderService.updateOrder(updatedOrder);
         return ResponseEntity.ok("Updated");
     }
 
-    @PatchMapping("/{id}")
-    @PreAuthorize("hasAuthority('TUTOR')")
-    public ResponseEntity<String> addTutorToOrder(@PathVariable("id") Long id,
-                                              @AuthenticationPrincipal Account account) {
+    public ResponseEntity<String> addTutorToOrder(Long id, Account account) {
         Tutor tutor = tutorService.findByAccount_Id(account.getId());
         Order order = orderService.findOrderById(id).orElseThrow(() -> new ResponseStatusException(
                 HttpStatus.NOT_FOUND, "Order not found"
@@ -80,21 +94,20 @@ public class OrderRestController {
         return ResponseEntity.ok("Updated");
     }
 
-    @GetMapping("/all")
-    @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<List<Order>> getAllOrders(@AuthenticationPrincipal Account account) {
+    public ResponseEntity<List<OrderDto>> getAllOrders(Account account) {
         Student student = studentService.findByAccount_Id(account.getId());
+        OrderConverter orderConverter = new OrderConverter(student);
         List<Order> orders = orderService.getOrdersByAuthor(student).orElseThrow(() -> new ResponseStatusException(
                 HttpStatus.NOT_FOUND, "Orders not found"
         ));
-        return ResponseEntity.of(Optional.of(orders));
+        List<OrderDto> orderDtoList = orders.stream()
+                .map(o -> (OrderDto) orderConverter.convert(o, TypeDescriptor.valueOf(Order.class), TypeDescriptor.valueOf(OrderDto.class))).collect(Collectors.toList());
+        return ResponseEntity.ok(orderDtoList);
     }
 
-    @GetMapping("/subjects")
-    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<List<Subject>> getAllSubjects() {
         List<Subject> subjects = subjectService.findAllSubjects();
-        return ResponseEntity.of(Optional.of(subjects));
+        return ResponseEntity.ok(subjects);
     }
